Extract draw logic into hook
